// src/app/Modules/core/models/User.ts
// Generated by https://quicktype.io

export interface User {
  username:        string;
  office_id:       string;
  rol_id:          string;
  user_type:       string;
  hashed_password: string;
  first_name:      string;
  last_name:       string;
  email:           string;
  phone:           string;
  status:          number;
  date_created:    string;
}

// Generated by https://quicktype.io

export interface UserToPost extends Omit<User, 'date_created' | 'hashed_password'> {
  password  : string,
  confirm : string,
}